+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmGenreStorage.java
@@ -1,10 +1,10 @@
-package ru.yandex.practicum.filmorate.dao;
+package ru.yandex.practicum.filmorate.storage;
 
 import ru.yandex.practicum.filmorate.model.Genre;
 
 import java.util.List;
 
-public interface FilmGenreDao {
+public interface FilmGenreStorage {
     void create(long filmId, long genreId);
 
     List<Genre> readGenresByFilmId(long filmId);
