+++ src/main/java/ru/yandex/practicum/filmorate/storage/FriendshipStorage.java
@@ -1,10 +1,10 @@
-package ru.yandex.practicum.filmorate.dao;
+package ru.yandex.practicum.filmorate.storage;
 
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
 
-public interface FriendshipDao {
+public interface FriendshipStorage {
 
     void create(long userId, long friendId);
