+++ src/main/java/ru/yandex/practicum/filmorate/storage/GenreStorage.java
@@ -1,10 +1,10 @@
-package ru.yandex.practicum.filmorate.dao;
+package ru.yandex.practicum.filmorate.storage;
 
 import ru.yandex.practicum.filmorate.model.Genre;
 
 import java.util.Collection;
 
-public interface GenreDao {
+public interface GenreStorage {
     Genre create(Genre genre);
 
     Collection<Genre> readAll();
