+++ src/main/java/ru/yandex/practicum/filmorate/storage/LikesStorage.java
@@ -1,11 +1,11 @@
-package ru.yandex.practicum.filmorate.dao;
+package ru.yandex.practicum.filmorate.storage;
 
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
 
-public interface LikesDao {
+public interface LikesStorage {
     void create(long userId, long filmId);
 
     List<User> readLikesByFilmId(long filmId);
