+++ src/main/java/ru/yandex/practicum/filmorate/storage/MpaStorage.java
@@ -1,10 +1,10 @@
-package ru.yandex.practicum.filmorate.dao;
+package ru.yandex.practicum.filmorate.storage;
 
 import ru.yandex.practicum.filmorate.model.Mpa;
 
 import java.util.List;
 
-public interface MpaDao {
+public interface MpaStorage {
     Mpa create(Mpa mpa);
 
     List<Mpa> readAll();
