+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.filmorate.storage.user;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.Collection;
-import java.util.Set;
 
 public interface UserStorage {
 
@@ -14,6 +13,4 @@ public interface UserStorage {
     public Collection<User> getUsers();
 
     public User getUserById(long userId);
-
-    public Collection<User> getUsersById(Set<Long> idCommonFriends);
 }
